backend/requirements.txt
langchain
langchain_community
langchain-pinecone
langchain-text-splitters
langchain-ollama
fastapi  
uvicorn[standard] 
langchain-experimental
python-dotenv

